Move form-summary to flat structure like other components
