+++ libs/core/src/components/form-summary/summary.component.ts
@@ -4,15 +4,15 @@ import { property, queryAsync } from 'lit/decorators.js'
 import { until } from 'lit/directives/until.js'
 import { when } from 'lit/directives/when.js'
 
-import { GdsElement } from '../../../gds-element'
-import { gdsCustomElement, html } from '../../../scoping'
-import { tokens } from '../../../tokens.style'
-import { GdsButton } from '../../button/button.component'
-import { GdsCard } from '../../card/card.component'
-import { GdsDiv } from '../../div/div.component'
-import { GdsFlex } from '../../flex/flex.component'
-import { IconArrowUp } from '../../icon/icons/arrow-up.component'
-import { GdsFormControlElement } from '../form-control'
+import { GdsElement } from '../../gds-element'
+import { gdsCustomElement, html } from '../../scoping'
+import { tokens } from '../../tokens.style'
+import { GdsButton } from '../button/button.component'
+import { GdsCard } from '../card/card.component'
+import { GdsDiv } from '../div/div.component'
+import { GdsFlex } from '../flex/flex.component'
+import { GdsFormControlElement } from '../form/form-control'
+import { IconArrowUp } from '../icon/icons/arrow-up.component'
 import styles from './summary.styles'
 
 /**
