+++ libs/core/src/components/form/index.ts
@@ -1,1 +1,1 @@
-export * from './summary'
+export * from '../form-summary'
