+++ libs/core/src/components/form/validation.mdx
@@ -1,6 +1,6 @@
 import { Meta, Markdown } from '@storybook/addon-docs'
 
-<Meta title="Components/Form/Validation/Documentation" />
+<Meta title="Components/Form/Validation" />
 
 <h1>Form validation</h1>
