+++ libs/core/src/components/pure.ts
@@ -31,7 +31,7 @@ export * from './icon/icon.component'
 export * from './segmented-control/segmented-control.component'
 export * from './theme/theme.component'
 export * from './filter-chips/filter-chips.component'
-export * from './form/summary/summary.component'
+export * from './form-summary/summary.component'
 export * from './input/input.component'
 export * from './select/select.component'
 export * from './spinner/spinner.component'
